Move everything to playback package
