+++ app/src/main/kotlin/com/simplemobiletools/musicplayer/playback/player/AudioOnlyRenderersFactory.kt
@@ -1,4 +1,4 @@
-package com.simplemobiletools.musicplayer.services.playback.player
+package com.simplemobiletools.musicplayer.playback.player
 
 import android.content.Context
 import android.os.Handler
